requirements.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.2
beautifulsoup4==4.4.1
itsdangerous==0.24
pyserial==2.7
python-dateutil==1.5
pytz==2015.7
requests==2.8.1
stravalib==0.5.0
units==0.06
wsgiref==0.1.2
